Fixes test setup for invoker tests
