+++ instrumentation/kafka-clients/src/it/kafka1/pom.xml
@@ -42,7 +42,7 @@
     <dependency>
       <groupId>com.github.charithe</groupId>
       <artifactId>kafka-junit</artifactId>
-      <version>@kafka-junit.version@</version>
+      <version>4.0.0</version>
     </dependency>
   </dependencies>
