+++ instrumentation/spring-web/src/it/spring3/pom.xml
@@ -35,6 +35,14 @@
       <scope>test</scope>
     </dependency>
 
+
+    <dependency>
+      <groupId>org.apache.httpcomponents</groupId>
+      <artifactId>httpclient</artifactId>
+      <version>@httpclient.version@</version>
+      <scope>test</scope>
+    </dependency>
+
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
