+++ instrumentation/spring-webmvc/src/test/java/brave/spring/webmvc/ITSpanCustomizingHandlerInterceptor.java
@@ -3,9 +3,15 @@ package brave.spring.webmvc;
 import java.util.EnumSet;
 import javax.servlet.DispatcherType;
 import org.eclipse.jetty.servlet.ServletContextHandler;
+import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
 
 public class ITSpanCustomizingHandlerInterceptor extends BaseITSpanCustomizingHandlerInterceptor {
   @Override protected void addDelegatingTracingFilter(ServletContextHandler handler) {
     handler.addFilter(DelegatingTracingFilter.class, "/*", EnumSet.allOf(DispatcherType.class));
+  }
+
+  @Override
+  protected void registerTestController(AnnotationConfigWebApplicationContext appContext) {
+    appContext.register(Servlet3TestController.class); // the test resource
   }
 }
