+++ instrumentation/spring-webmvc/src/test/java/brave/spring/webmvc/Servlet25TestController.java
@@ -5,7 +5,6 @@ import brave.http.HttpTracing;
 import brave.propagation.ExtraFieldPropagation;
 import brave.test.http.ITHttp;
 import java.io.IOException;
-import java.util.concurrent.Callable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,10 +13,10 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-@Controller class TestController {
+@Controller class Servlet25TestController {
   final Tracer tracer;
 
-  @Autowired TestController(HttpTracing httpTracing) {
+  @Autowired Servlet25TestController(HttpTracing httpTracing) {
     this.tracer = httpTracing.tracing().tracer();
   }
 
@@ -47,31 +46,14 @@ import org.springframework.web.bind.annotation.RequestMethod;
     return new ResponseEntity<>(HttpStatus.OK);
   }
 
-  @RequestMapping(value = "/async")
-  public Callable<ResponseEntity<Void>> async() {
-    return () -> new ResponseEntity<>(HttpStatus.OK);
-  }
-
   @RequestMapping(value = "/exception")
   public ResponseEntity<Void> disconnect() throws IOException {
     throw new IOException();
   }
 
-  @RequestMapping(value = "/exceptionAsync")
-  public Callable<ResponseEntity<Void>> disconnectAsync() {
-    return () -> {
-      throw new IOException();
-    };
-  }
-
   @RequestMapping(value = "/items/{itemId}")
   public ResponseEntity<String> items(@PathVariable("itemId") String itemId) {
     return new ResponseEntity<String>(itemId, HttpStatus.OK);
-  }
-
-  @RequestMapping(value = "/async_items/{itemId}")
-  public Callable<ResponseEntity<String>> asyncItems(@PathVariable("itemId") String itemId) {
-    return () -> new ResponseEntity<String>(itemId, HttpStatus.OK);
   }
 
   @Controller
