+++ pom.xml
@@ -84,9 +84,12 @@
     <finagle.version>18.7.0</finagle.version>
     <log4j.version>2.11.0</log4j.version>
     <okhttp.version>3.11.0</okhttp.version>
+    <httpclient.version>4.5.6</httpclient.version>
+
     <grpc.version>1.14.0</grpc.version>
     <!-- prefer grpc's version of netty -->
     <netty.version>4.1.27.Final</netty.version>
+
     <sparkjava.version>2.7.1</sparkjava.version>
     <junit.version>4.12</junit.version>
     <assertj.version>3.11.0</assertj.version>
@@ -205,12 +208,12 @@
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
             <artifactId>httpclient</artifactId>
-            <version>4.5.6</version>
+            <version>${httpclient.version}</version>
         </dependency>
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
             <artifactId>httpclient-cache</artifactId>
-            <version>4.5.6</version>
+            <version>${httpclient.version}</version>
         </dependency>
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
